+++ includes/class-rest-reports.php
@@ -95,6 +95,7 @@ final class RestReports {
 		return Phase3FeatureSettings::enabled( 'reports_enabled' )
 			&& function_exists( 'is_user_logged_in' )
 			&& is_user_logged_in()
+			&& CanonicalIdentityAdapter::current_action_ready( (int) get_current_user_id() )
 			&& InteractionPermissions::nonce_valid( self::request_nonce( $request ) );
 	}
 
@@ -107,6 +108,7 @@ final class RestReports {
 	public static function moderator_permission( $request ) {
 		return function_exists( 'is_user_logged_in' )
 			&& is_user_logged_in()
+			&& CanonicalIdentityAdapter::current_action_ready( (int) get_current_user_id() )
 			&& InteractionPermissions::nonce_valid( self::request_nonce( $request ) )
 			&& InteractionPermissions::can_manage_reports();
 	}